make HealthCards non-collapsible, change badges
