+++ packages/front-end/components/HealthTab/StatusBadge.tsx
@@ -1,37 +1,18 @@
-import Tooltip from "../Tooltip/Tooltip";
-
 export const BadgeColors = {
-  healthy: "badge-green",
-  unhealthy: "badge-red",
-  "not enough data": "badge-gray",
+  "Issues detected": "badge-warning",
+  "Not enough traffic": "badge-light",
 };
-export type HealthStatus = keyof typeof BadgeColors;
+export type HealthStatus = keyof typeof BadgeColors | "healthy";
 
 interface Props {
   status: HealthStatus;
-  hasTooltip?: boolean;
-  tooltipBody?: string | JSX.Element;
 }
 
-export const StatusBadge = ({
-  status,
-  hasTooltip = true,
-  tooltipBody = "",
-}: Props) => {
-  if (hasTooltip) {
-    return (
-      <Tooltip
-        body={tooltipBody}
-        className={"badge border ml-2 mr-2 " + BadgeColors[status]}
-        tipPosition="top"
-      >
-        {status}
-      </Tooltip>
-    );
-  }
-
+export const StatusBadge = ({ status }: Props) => {
   return (
-    <span className={"badge border ml-2 mr-2 " + BadgeColors[status]}>
+    <span
+      className={"badge badge-pill border ml-2 mr-2 " + BadgeColors[status]}
+    >
       {status}
     </span>
   );
